requirements.txt
astronomer-cosmos>=1.3.2
dbt-core>=1.7.9
dbt-athena-community>=1.7.1
dbt-postgres>=1.7.0
awswrangler>=3.8.0
pandas